requirements.txt
paho-mqtt
pyinstaller
py2app